refactor: Arrange file structure
