+++ src/processor/theme.ts
@@ -3,8 +3,8 @@ import fs from 'node:fs';
 import npa from 'npm-package-arg';
 import shelljs from 'shelljs';
 import path from 'upath';
-import { MergedConfig } from './config.js';
-import { beforeExitHandlers, DetailError } from './util.js';
+import type { MergedConfig } from '../input/config.js';
+import { beforeExitHandlers, DetailError } from '../util.js';
 
 // Rename `packages` directory into `node_modules` while Arborist works
 const temporaryMovePackagesDirectrory = async <T = unknown>(
